Fix torture tests: dead code, large types, float cast, non-local goto — 974/995 (97.9%)
- Dead code elimination for if(0) inside switch: skip unreachable code
  but preserve case labels (gen_stmt_dead helper) — fixes medce-1
- Widen Type.size, Member.offset, array_len, align_to to long for huge
  struct/union sizeof correctness — fixes 991014-1
- Float-to-int: use fcvtzs/fcvtzu w0 for 32-bit targets so ARM64
  saturates properly at INT_MAX/INT_MIN — fixes 20031003-1
- Non-local goto from nested functions via longjmp: set up [fp, sp,
  label_addr] buffer in outer function prologue, longjmp from nested
  function through static chain — fixes 920428-2, 920501-7, comp-goto-2
- Fix nested function recursive self-calls: push function name into
  scope before parsing body so recursive calls resolve to mangled name
- Large struct member offset codegen: use load_imm for offsets > 4095

+++ src/type.c
@@ -119,7 +119,7 @@ Type *func_type(Type *return_ty) {
   return ty;
 }
 
-Type *array_of(Type *base, int len) {
+Type *array_of(Type *base, long len) {
   Type *ty = calloc_checked(1, sizeof(Type));
   ty->kind = TY_ARRAY;
   ty->size = base->size * len;
